+++ Core/Module/CharacterData/Template/ITemplateHandler.cs
@@ -4,12 +4,12 @@
     {
         byte GetClassId();
         byte GetRaceId();
-        int GetInt();
-        int GetStr();
-        int GetCon();
-        int GetMen();
-        int GetDex();
-        int GetWit();
+        byte GetInt();
+        byte GetStr();
+        byte GetCon();
+        byte GetMen();
+        byte GetDex();
+        byte GetWit();
         float GetCpBegin(int level);
         float GetHpBegin(int level);
         float GetMpBegin(int level);
